Master a corrigé le cast
cast ajouté

+++ WindowsFormsApplication29/Form1.cs
@@ -18,8 +18,9 @@ namespace WindowsFormsApplication29
 
         private void Form1_Load(object sender, EventArgs e)
         {
+            // Master : je vais modifié cela
             int? i = null;
-            int j = 512 + i;
+            int j = 512 + (int)i;
         }
     }
 }
